requirement.txt
ortools==9.12.4544
psycopg2==2.9.10
python-dotenv==1.1.0
uvicorn==0.34.0
fastapi==0.115.12